Emulator: Now boots tetris until the copyright scr
Some glitches in the font though, probably one or two instructions not setting the flags properly.
Soft reset does work, as the long as the keys are pressed since the beginning, since there are no interrupts yet.
All of CPU done and integrated, now needs bug fixes and correction in the instructions. Also needs the integration in the opcode table

Coming next -> Use of templates to quickly implement CPU functions and generate one for each parameter (to reduce calls and improve speed?)
	       Cleaning of all the small things in the code and complete integration of the new CPU
	       Checking of all the instruction set using the blargg test roms.